backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
openai==1.54.0
psycopg[binary,pool]==3.2.2
ddtrace==2.14.0
python-dotenv==1.0.1
pydantic==2.9.0
wrapt==1.16.0
httpx==0.27.2
python-json-logger==2.0.7

